+++ bench/cache_freecache.go
